chore(cleanup): PR6 — collapse 3 duplicate type pairs (cluster H mechanical)
Closes 3 of 5 cluster-H type-overlap pairs from the hard-cutover triage.
The remaining 2 pairs need an architectural cutover and are intentionally
out of scope here — see "Descoped" below.

1. ISerializedUnitEnvelope (×2) — single generic
   `ISerializedUnitEnvelope<T = ISerializedUnit>` in
   `src/types/unit/UnitSerialization.ts`. Persistence keeps a dedicated
   alias `ISerializedUnitEnvelope = ISerializedUnitEnvelopeGeneric<Record<string, unknown>>`
   because `src/types/index.ts` namespaces the persistence barrel
   (`export * as persistence`) to avoid collision with the typed `unit`
   re-exports at the top-level barrel.

2. IAmmoConstruction{Init,Data} — extracted to new
   `src/types/gameplay/AmmoTypes.ts` (single source-of-truth). Both old
   locations now re-export from the new leaf:
   - `IAmmoConstructionInit` (in `GameSessionInterfaces.ts`) → re-export
     of the same shape for legacy import paths.
   - `IAmmoConstructionData` (in `utils/gameplay/ammoTracking/types.ts`)
     → re-export of the same shape for ammo-tracking call-sites.
   The cycle the duplicate guarded against doesn't actually exist — the
   types package never imports from utils/gameplay (verified via grep).

3. Dead `IUnit` stub deleted — `src/types/core/UnitInterfaces.ts` was a
   `@deprecated` stub re-exporting `LAMMode`/`QuadVeeMode` from
   `MechConfigurationSystem`. Confirmed via grep that no production
   code imports from `@/types/core/UnitInterfaces`; the only
   `MountedBattleArmor` reference outside the file is an unrelated
   Storybook story name. The `core` barrel doesn't re-export it. File
   deleted; the two enum re-exports are still available at their
   canonical path `@/types/construction/MechConfigurationSystem`.

DESCOPED (out of scope for PR6, need architecture council):

a. IFullUnit replacement — the brief proposed swapping
   `IFullUnit` (in `src/services/units/CanonicalUnitService.ts`) for the
   "discriminated union from BaseUnitInterfaces.ts." Reality:
   `BaseUnitInterfaces.ts` exposes a HIERARCHY (`IBaseUnit` →
   `IGroundUnit` / `IAerospaceUnit` / `ISquadUnit`) plus type-guard
   functions, NOT a discriminated `BattleMech | Vehicle | Aerospace |
   ...` union. `IFullUnit` is the load-bearing wire/IndexedDB shape
   with `[key: string]: unknown` index signature consumed by ~50 sites
   (`CustomUnitService`, `CompendiumAdapter`, `useUnitCardActions`,
   `unitContractAdapter`, `MechBuilderService`, etc.) that depend on
   the loose access pattern. Replacing it is a wire-format /
   serialization-boundary cutover, not a mechanical merge — belongs
   in a dedicated PR with explicit migration of every consumer.

b. IUnitWithVersion — `IUnitWithVersion = IFullUnit & {_version, ...}`
   is structurally exactly what `IFullUnit & IVersionedMeta` would be,
   but `IVersionedMeta` doesn't exist yet in the type tree. Collapsing
   this depends on (a) landing first.

Verification:
- npx tsc --noEmit --skipLibCheck — exit 0
- npx jest — 23166/23166 pass (881 suites, 1 preexisting skip)
- npx tsx scripts/validate-bv.ts — 389/389 calculated, 0 failed
  (numbers match pre-PR6 baseline; PR6 introduced ZERO BV drift).
- npx oxfmt --check . — clean (3227 files)
- npx oxlint — 33 baseline warnings, 0 errors (no NEW lint hits;
  the 1 max-lines warning on GameSessionInterfaces.ts is preexisting
  and the file is now SHORTER after the collapse).
- npm run analyze:dependencies — 1 preexisting cycle in
  types/campaign/CampaignInterfaces; PR6 introduced ZERO new cycles.

Net diff: -55 lines (95 deletions, 40 insertions across 7 files).

Part of the hard-cutover triage. PR1+PR2+PR3+PR4 already merged;
PR5 (singleton codemod) running in parallel.
